# api/openapi.yaml
openapi: 3.0.3
info:
  title: identity
  version: "1"
  description: |
    Who somebody is, once, for every service operated together. Sign in with
    GitHub or Discord, get one account, carry it everywhere.

    This file is the contract. `internal/api` implements it and a test holds
    the two equal: every route is described here, every path here is a route,
    and every response the test suite provokes is validated against these
    schemas. The Go client in this package is generated from it; a TypeScript
    client would be generated from the same file.

    Two kinds of credential exist and the difference is the security model.
    An **account token** comes from signing in here (or from `POST /v1/tokens`)
    and may manage the account. An **application token** comes from
    `POST /v1/exchange` and is bound to one application's origin (its
    `token.audience`); it may call `/v1/whoami` and `/v1/avatar`, revoke
    itself, sign its browser out with `/v1/signout`, and nothing else. A
    browser-facing service must hold only its own application tokens,
    server-side, and must check `token.audience` against its own origin
    before trusting a `whoami` answer.

    An application token minted for a browser that signed in here is linked
    to that sign-in: revoking the sign-in revokes the token, and signing in to
    the same application again from the same sign-in replaces it.
servers:
  - url: /
...
security:
  - bearer: []
paths:
...
  /v1/tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: The caller's live and recently revoked tokens.
      responses:
        "200":
          description: Tokens, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items: { $ref: "#/components/schemas/TokenInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Failure" }
    post:
      operationId: mintToken
      tags: [tokens]
      summary: Mint a named account token.
      description: |
        At most 25 account tokens are live at once; application tokens do not
        count toward it.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        "201":
          description: The token, shown once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Token" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/tokens/{id}:
    delete:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoke one of the caller's tokens.
      description: |
        Revoking a token revokes every application token minted from it. An
        application token may revoke only itself.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Failure" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer bsid_...`"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "What went wrong, in a sentence." }
    AccountID:
      type: string
      description: A random 12-character hex id. Stable for the life of the account.
      pattern: "^[0-9a-f]{12}$"
      example: 3f9c2a81d04b
...
    Token:
      type: object
      description: A freshly minted credential. The token itself is shown exactly once.
      required: [token, account, handle]
      properties:
        token: { type: string, description: "`bsid_<id>_<secret>`. Empty when set as a cookie instead." }
        account: { $ref: "#/components/schemas/AccountID" }
        handle: { type: string }
        expires_at: { type: string, format: date-time }
    TokenInfo:
      type: object
      description: A token as described, never including its secret.
      required: [id, name, audience, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        audience:
          type: string
          description: The application origin this token is bound to; empty for an account token.
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        revoked: { type: boolean }
        current: { type: boolean, description: This is the token making the request. }
...
  responses:
...
    Unauthorized:
      description: No live bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The token exists but may not do this; the error says why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Failure:
      description: This service could not do it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...